+++ src/openFrameworks/ofMath.cpp
@@ -30,13 +30,13 @@ void ofSeedRandom() {
 		struct timeval tv;
 		gettimeofday(&tv, 0);
 		long int n = (tv.tv_sec ^ tv.tv_usec) ^ getpid();
-		srand(n);
+		srand((unsigned int)n);
 	#endif
 }
 
 //--------------------------------------------------
 void ofSeedRandom(int val) {
-	srand((long) val);
+	srand((unsigned int) val);
 }
 
 //--------------------------------------------------
